Reserve VMEM_MINFREE for the vmem_seg_arena. When this arena needs to grow, we are in the populate mode and can only take segments from the global list. If we fail, we may fail allocation.
This should be safe because the worst case can happen is we are going to waste MINFREE
or 8 segments.  I did my four way tests for > 1 hour.  I also did snapshot when slash2
is running without seeing any problem.  With gdb, I did see vmem_nsegfree_reserve bumps
to over 8000. So the code is exercised.